# docs/03_REFERENCE/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "SOS Pereira API",
    "description": "Contrato para coordinación humanitaria territorial.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/reports": {
      "post": {
        "tags": [
          "reports"
        ],
        "summary": "Create Report",
        "operationId": "create_report_api_v1_reports_post",
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 16,
              "maxLength": 128,
              "title": "Idempotency-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReportCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReportRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/reports/public": {
      "get": {
        "tags": [
          "reports"
        ],
        "summary": "Public Reports",
        "operationId": "public_reports_api_v1_reports_public_get",
        "parameters": [
          {
            "name": "territory_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 3,
              "maxLength": 80,
              "title": "Territory Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PublicReportRead"
                  },
                  "title": "Response Public Reports Api V1 Reports Public Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/reports/moderation/pending": {
      "get": {
        "tags": [
          "reports"
        ],
        "summary": "Pending Reports",
        "operationId": "pending_reports_api_v1_reports_moderation_pending_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ReportRead"
                  },
                  "type": "array",
                  "title": "Response Pending Reports Api V1 Reports Moderation Pending Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/reports/moderation/{report_id}": {
      "patch": {
        "tags": [
          "reports"
        ],
        "summary": "Moderate Report",
        "operationId": "moderate_report_api_v1_reports_moderation__report_id__patch",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "report_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Report Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReportModerationUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReportStatusRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/reports/{code}/status": {
      "get": {
        "tags": [
          "reports"
        ],
        "summary": "Report Status",
        "operationId": "report_status_api_v1_reports__code__status_get",
        "parameters": [
          {
            "name": "code",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Code"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReportStatusRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Coordinates": {
        "properties": {
          "longitude": {
            "type": "number",
            "maximum": 180.0,
            "minimum": -180.0,
            "title": "Longitude"
          },
          "latitude": {
            "type": "number",
            "maximum": 90.0,
            "minimum": -90.0,
            "title": "Latitude"
          }
        },
        "type": "object",
        "required": [
          "longitude",
          "latitude"
        ],
        "title": "Coordinates"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PublicReportRead": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "territory_id": {
            "type": "string",
            "title": "Territory Id"
          },
          "category": {
            "$ref": "#/components/schemas/ReportCategory"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "description": {
            "type": "string",
            "title": "Description"
          },
          "neighborhood": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Neighborhood"
          },
          "severity": {
            "$ref": "#/components/schemas/Severity"
          },
          "coordinates": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Coordinates"
              },
              {
                "type": "null"
              }
            ]
          },
          "observed_at": {
            "type": "string",
            "format": "date-time",
            "title": "Observed At"
          },
          "verification_status": {
            "$ref": "#/components/schemas/VerificationStatus"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "territory_id",
          "category",
          "title",
          "description",
          "neighborhood",
          "severity",
          "coordinates",
          "observed_at",
          "verification_status",
          "updated_at"
        ],
        "title": "PublicReportRead",
        "description": "Privacy-preserving projection for the public map."
      },
...
      "ReportCategory": {
        "type": "string",
        "enum": [
          "need",
          "offer",
          "shelter",
          "aid-center",
          "damage"
        ],
        "title": "ReportCategory"
      },
      "ReportCreate": {
        "properties": {
          "territory_id": {
            "type": "string",
            "maxLength": 80,
            "minLength": 3,
            "title": "Territory Id",
            "default": "co-ris-pereira"
          },
          "category": {
            "$ref": "#/components/schemas/ReportCategory"
          },
          "title": {
            "type": "string",
            "maxLength": 100,
            "minLength": 3,
            "title": "Title"
          },
          "description": {
            "type": "string",
            "maxLength": 600,
            "minLength": 5,
            "title": "Description"
          },
          "neighborhood_code": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 50
              },
              {
                "type": "null"
              }
            ],
            "title": "Neighborhood Code"
          },
          "severity": {
            "$ref": "#/components/schemas/Severity"
          },
          "coordinates": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Coordinates"
              },
              {
                "type": "null"
              }
            ]
          },
          "observed_at": {
            "type": "string",
            "format": "date-time",
            "title": "Observed At"
          },
          "privacy_authorized": {
            "type": "boolean",
            "const": true,
            "title": "Privacy Authorized"
          },
          "privacy_policy_version": {
            "type": "string",
            "maxLength": 32,
            "minLength": 1,
            "title": "Privacy Policy Version"
          }
        },
        "type": "object",
        "required": [
          "category",
          "title",
          "description",
          "severity",
          "observed_at",
          "privacy_authorized",
          "privacy_policy_version"
        ],
        "title": "ReportCreate"
      },
      "ReportModerationUpdate": {
        "properties": {
          "verification_status": {
            "$ref": "#/components/schemas/VerificationStatus"
          },
          "moderation_note": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 300
              },
              {
                "type": "null"
              }
            ],
            "title": "Moderation Note"
          }
        },
        "type": "object",
        "required": [
          "verification_status"
        ],
        "title": "ReportModerationUpdate"
      },
      "ReportRead": {
        "properties": {
          "territory_id": {
            "type": "string",
            "maxLength": 80,
            "minLength": 3,
            "title": "Territory Id",
            "default": "co-ris-pereira"
          },
          "category": {
            "$ref": "#/components/schemas/ReportCategory"
          },
          "title": {
            "type": "string",
            "maxLength": 100,
            "minLength": 3,
            "title": "Title"
          },
          "description": {
            "type": "string",
            "maxLength": 600,
            "minLength": 5,
            "title": "Description"
          },
          "neighborhood_code": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 50
              },
              {
                "type": "null"
              }
            ],
            "title": "Neighborhood Code"
          },
          "severity": {
            "$ref": "#/components/schemas/Severity"
          },
          "coordinates": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Coordinates"
              },
              {
                "type": "null"
              }
            ]
          },
          "observed_at": {
            "type": "string",
            "format": "date-time",
            "title": "Observed At"
          },
          "privacy_authorized": {
            "type": "boolean",
            "const": true,
            "title": "Privacy Authorized"
          },
          "privacy_policy_version": {
            "type": "string",
            "maxLength": 32,
            "minLength": 1,
            "title": "Privacy Policy Version"
          },
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "tracking_code": {
            "type": "string",
            "title": "Tracking Code"
          },
          "verification_status": {
            "$ref": "#/components/schemas/VerificationStatus"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "category",
          "title",
          "description",
          "severity",
          "observed_at",
          "privacy_authorized",
          "privacy_policy_version",
          "id",
          "tracking_code",
          "verification_status",
          "created_at"
        ],
        "title": "ReportRead"
      },
      "ReportStatusRead": {
        "properties": {
          "tracking_code": {
            "type": "string",
            "title": "Tracking Code"
          },
          "verification_status": {
            "$ref": "#/components/schemas/VerificationStatus"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "tracking_code",
          "verification_status",
          "created_at",
          "updated_at"
        ],
        "title": "ReportStatusRead"
      },
...
      "Severity": {
        "type": "string",
        "enum": [
          "low",
          "medium",
          "high",
          "critical"
        ],
        "title": "Severity"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
      "VerificationStatus": {
        "type": "string",
        "enum": [
          "reported",
          "verified",
          "official",
          "stale",
          "closed",
          "rejected"
        ],
        "title": "VerificationStatus"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}